refactor global firestore data
