+++ src/global-firestore-data/toggleable-requirement-choices.ts
@@ -1,0 +1,13 @@
+import { toggleableRequirementChoicesCollection } from '@/firebase-frontend-config';
+import store from '@/store';
+
+// TODO: add functions addToggleableRequirementChoice and deleteToggleRequirementChoice
+const chooseToggleableRequirementOption = (
+  toggleableRequirementChoices: AppToggleableRequirementChoices
+): void => {
+  toggleableRequirementChoicesCollection
+    .doc(store.state.currentFirebaseUser.email)
+    .set(toggleableRequirementChoices);
+};
+
+export default chooseToggleableRequirementOption;
